Add i18n:extract npm script
...instead of doing it in all production builds
