+++ package.json
@@ -29,6 +29,7 @@
     "ensure-config-exists": "node ./bin/ensure-config-exists.js",
     "ensure-uploads-dir-exists": "node ./bin/ensure-uploads-dir-exists.js",
     "fontello": "gulp fontello",
+    "i18n:extract": "babel -f babel.config.js -o /dev/null 'modules/**/*.js'",
     "install-dependencies": "node ./bin/install-deps.js",
     "lint:watch": "esw --cache --watch --quiet .",
     "lint": "eslint .",
@@ -182,6 +183,7 @@
     "winston-papertrail": "1.0.5"
   },
   "devDependencies": {
+    "@babel/cli": "7.8.4",
     "@babel/core": "7.9.6",
     "@babel/plugin-proposal-nullish-coalescing-operator": "7.8.3",
     "@babel/plugin-proposal-object-rest-spread": "7.9.6",
